resources/requirements.txt
GitPython
pytest
requests